reducing number of test cases with ONE_PER_TYPE
