+++ cfme/tests/v2v/test_csv_import.py
@@ -7,7 +7,7 @@ from widgetastic.exceptions import UnexpectedAlertPresentException
 from cfme.fixtures.v2v import _form_data
 from cfme.infrastructure.provider.rhevm import RHEVMProvider
 from cfme.infrastructure.provider.virtualcenter import VMwareProvider
-from cfme.markers.env_markers.provider import ONE_PER_VERSION
+from cfme.markers.env_markers.provider import ONE_PER_VERSION, ONE_PER_TYPE
 from cfme.utils.appliance.implementations.ui import navigate_to
 from cfme.utils.blockers import BZ
 from cfme.utils.generators import random_vm_name
@@ -20,7 +20,7 @@ pytestmark = [
     ),
     pytest.mark.provider(
         classes=[VMwareProvider],
-        selector=ONE_PER_VERSION,
+        selector=ONE_PER_TYPE,
         fixture_name='second_provider'
     )
 ]
