+++ cfme/tests/v2v/test_v2v_ansible.py
@@ -5,7 +5,7 @@ from cfme import test_requirements
 from cfme.fixtures.provider import rhel7_minimal
 from cfme.infrastructure.provider.rhevm import RHEVMProvider
 from cfme.infrastructure.provider.virtualcenter import VMwareProvider
-from cfme.markers.env_markers.provider import ONE_PER_VERSION
+from cfme.markers.env_markers.provider import ONE_PER_VERSION, ONE_PER_TYPE
 from cfme.utils.appliance.implementations.ui import navigate_to, navigator
 from cfme.utils.conf import cfme_data, credentials
 from cfme.utils.log import logger
@@ -24,7 +24,7 @@ pytestmark = [
     ),
     pytest.mark.provider(
         classes=[VMwareProvider],
-        selector=ONE_PER_VERSION,
+        selector=ONE_PER_TYPE,
         fixture_name="second_provider",
         required_flags=["v2v"]
     ),
